feat: RevenueCat + fix Riverpod v3 .g.dart files
- Fix quiz_provider.g.dart: rewrite for Riverpod v3 API
  ($FunctionalProvider, $NotifierProvider, _$QuizNotifier base class)
- Fix market_data_provider.g.dart: same v3 rewrite
- Fix education_home_screen.dart:109: num → int cast
- feat(subscription): full RevenueCat integration
  - main.dart: Purchases.configure() on mobile startup
  - subscription_provider.dart: purchase(), restore(), customerInfoStream
  - subscription_screen.dart: real packages from offerings, loading states,
    restore purchases flow, web fallback, graceful empty state

+++ lib/features/education/presentation/screens/education_home_screen.dart
@@ -106,7 +106,7 @@ class EducationHomeScreen extends ConsumerWidget {
                       quiz: quiz,
                       lastAttempt: lastAttempt,
                       onTap: () => context.push('${AppRoutes.education}/quiz/${quiz.id}'),
-                    ).animate().fadeIn(delay: Duration(milliseconds: 360 + e.key * 60));
+                    ).animate().fadeIn(delay: Duration(milliseconds: (360 + e.key * 60).toInt()));
                   }).toList(),
                 );
               },
